+++ ObjectPositions.h
@@ -60,8 +60,8 @@ namespace brightland {
 		float CalcObjectXPositionPercentage(float position, std::string objectName);
 		float CalcObjectXAbsolutePosition(float position, std::string objectName);
 		float CalcObjectYAbsolutePosition(float position, std::string objectName);
-		float CalcAssetSize(int size,int percentIncrease);
-		float CalcAssetDecreaseSize(int size, int percentDecrease);
+		int CalcAssetSize(int size,int percentIncrease);
+		int CalcAssetDecreaseSize(int size, int percentDecrease);
 
 		void PlaceMenuYPosition();
 		void PlaceHoverObjects();
